added initial auth functionality based on authlogic. (revised)
At the moment you've got to install the authlogic gem -- we will eventually need to package it as an RPM and add it as a dependency in the specfile.

Right now any user can create an account and access all pages. We aren't doing any access control beyond requiring a login account.