first round of model changes for new cloud aggregator design
This is part of an overall redesign of the codebase. The most significant change
is that a pool is now made up of one or more "back end" (or provider) accounts.
Initially a user will have to pick a back end account (via the selection of a
realm), but ultimately provider account selection will be optional, and the
scheduler will automatically choose a back end account for each instance.
In addition resources such as images or hardware profiles will be mapped from
front end to back end clouds so that pool-specific images and hardware profiles
chosen by the user will be mapped to the appropriate resources for whichever
back-end cloud account ultimately receives the 'new instance' request.

1) the full hardware profile model is not included -- this is roughly the old flavor model renamed to 'hardware profile'
2) images and hardware profiles do not reflect the ability to map back-end images and profiles of multiple architectures to a single front end image or profile.