Rails version and db config updates.
This patch updates environment so portal will run on both F11 and F12
as well as adding sample config files to support mysql and sqlite. If
a developer/admin wants to use either, they can just rename them to
database.yml.  Perhaps we can do the same thing with postgres and have no
versioned .yml file, always leave it to person installing to choose their
db of choice, but if we do that, it will be in a subsequent patch.