Add infrastructure for BDD development.
This patch adds Rails integration points for using
cucumber, rspec and webrat, first batch of scenarios/specs
will be in a separate upcoming patch.

Note for fedora that rspec and cucumber are available via
yum, but cucumber currently has an incorrect version for
one of its dependencies, so for the time being it is probably
best to just gem install everything needed besides rspec.