Port controller tests to rspec, route & dep changes.
There are a few non-obvious changes to routes, and an extra
testing library (factory_girl).  As a result, we can now (but
don't _have to_) use factories instead of yaml files for test
data.  The idea here is to make things more readable, as well as
remove duplication, since one Factory can inherit from another.
Each still gets it's own object-named file though, so the structure
is similar to what we have done in the past.  See
http://github.com/thoughtbot/factory_girl for more details.

All existing TestUnit tests should still run (with rake test), and
the new rspec tests can be run in a similar manner with:
rake spec