Initial set of Cucumber features.
This covers just auth-related app-level features, atm.
I have it set up now, so when anyone else writes a new feature
or scenario that requires login, they can reuse the bits in here
(as in, reference them not duplicate).  There may be some tweaking
needed for that as we go forward, but this should be a good starting
point.  Eventually we will probably use cucmber features in place of
most or all of the rspec controller tests, as that is a current area
of duplication.

Tests can be run from the src directory with either:
cucumber
 or
rake cucumber