Add initial integration with Image Builder.
This is in the form of a script that could be run as a service.
For now it needs to be run in the foreground.

Note that this is highly experimental, and requires an
Image Builder console gem be installed (which should be publicly
available soon).

Run the script with:
  ruby runner/image_builder_service.rb

Logging needs cleanup, right now it is more or less on DEBUG, once
this is past experimental, that will be defaulted to WARN.