Added provider account matchmaking to current task scheduler.
Note that this is not part of the condor infrastructure -- it's an interim solution to handle scheduling across providers until the full scheduling capabilities are put into use with Condor integration.

The full list of provider accounts is filtered by realm, hardware profile, and image -- so only those accounts that map validly to the "front-end/engine" realm/hwp/image choices of the user are considered. After this, we filter out those cloud accounts that do not have the available capacity to schedule the instance creation request.

As we do not  yet have a UI to map images and HWPs from different providers to the same front-end image or HWP, multi-provider scheduling is only possible across accounts in the same provider, or across providers after manually mapping equivalent images and HWPs.

For a simple test case with exactly two providers in the AR database with the same set of available images and HWPs (this will only be true for two mock providers, or if you've set up two separate provider objects pointing to the same driver -- or two drivers pointing to the same cloud impl), the following should take your separately-mapped resources and map them together on the front end:

Provider.last.images.each do  |image|
  if image.aggregator_images[0].provider_images.size < 2
    image.aggregator_images[0].delete
    image.aggregator_images = [Provider.first.images.find_by_external_key(image.external_key).aggregator_images[0]]
    image.save
  end
  agg_image = image.aggregator_images[0]
  agg_image.name = image.name
  agg_image.external_key = image.external_key
  agg_image.save
end

Provider.last.hardware_profiles.each do  |hardware_profile|
  if hardware_profile.aggregator_hardware_profiles[0].provider_hardware_profiles.size < 2
    hardware_profile.aggregator_hardware_profiles[0].delete
    hardware_profile.aggregator_hardware_profiles = [Provider.first.hardware_profiles.find_by_external_key(hardware_profile.external_key).aggregator_hardware_profiles[0]]
    hardware_profile.save
  end
  agg_hwp = hardware_profile.aggregator_hardware_profiles[0]
  agg_hwp.name = hardware_profile.name
  agg_hwp.external_key = hardware_profile.external_key
  agg_hwp.save
end