First cut at Condor to implement jobs (revision 3)
This patch is a first try at using condor as a job management system.
This removes the usage of the 'taskomatic' utilities and replaces them
with 'condormatic' calls that use the command line interfaces (no qmf or
gsoap etc) to condor.

On startup of the server (and any changes after running), a pile of
'classads' are created which define each possible startup location for
a given set of image/hardware profiles that exist and are useable as
well as the backend info condor needs to start an instance on the given
provider.

For each instance that you start, a job will be created in condor.
Condor will then match the hardware profile and image to a provider and
can then start an instance on that provider.  When you stop or destroy
that instance, the job will be removed (which isn't really how we want
it to go but..).

This patch requires that you have our custom hacked up condor installed.
You can get this at:

http://people.redhat.com/clalance/condor-dcloud

Be sure to read the README.  Chris has written up very good instructions
on how to set up condor.

REVISED: This patch plus the new condor fixes a number of the bugs that
were in the previous patch.  This patch adds realm matching support and
fixes the start/stop issues we were seeing.  So most things basically
work now and I think it's generally useable.  Probably the biggest
outstanding bug for useability is that we do not keep long-running jobs
for stateful instances.

The outstanding bugs are now limited to:

- To 'stop' a job in condor we should be using 'hold' instead of
  removing the job.  This is creating a few different problems.
- We are still reaching directly to the DeltaCloud API to get a list of
  available actions for each instance.  Maybe this is fine, I'm not
  sure.
- Quotas are not yet implemented.
- Classads are sync'd to condor on startup and on any changes to the
  hardware profile and image records.  However, if you restart condor
  you won't have any classads in it to match against and your jobs will
  fail.
- We're still using 'on-demand' syncing of states from condor to the
  aggregator.  eg when you list the instances it updates the states of
  each instance from condor at that time.  There is no event logging.
- There's no 'reboot' as yet in condor.  Not sure how we'll deal with
  that just yet.
- We've kept the tasks model and usage but they are quazi-meaningless.
  The task table needs to turn into an event/audit log table.

Signed-off-by: Ian Main <imain@redhat.com>