Unhook the cucumber tests from a running core server.
We may want (and probably do) a truly full stack test
that has all external components running, but that is not
the goal with the cucumber tests.  External dependencies
should be tested in the 3rd-party library, so for our
acceptance testing, just assume those kinds of things
work as advertised.

Again, a full 'integrated' test suite should probably be
incorporated at some point as well.