Remove the left sidebar.
It doesn't make sense to display the sidebar all the time -- even when
nobody's logged in. This removes it from most of the pages.

To select a specific pool, go to Instances in the main menu.

To see a list of providers, go to the Settings menu.

Note that the layout is likely to change in the near future.