Reworked image/instance listing
Replaced Datatables plugin with simpler HTML table because some
issues when JS is off. Major issue with Datatables is that it doesn't
degrades gracefully when JS is off and server-side pagination.

Images and instances are now displayed by HTML table
(there is new helper 'paginated_table'), this table keeps
all common functions like sorting, paginating, searching and
code is cleaner. Works with JS on or off.