Make realm NULL if not specified by the user.
Our classad matching was using the realm from the provider classad even
if the user didn't specify it; it was just matching against the first
provider classad regardless of realm and using that.

This patch makes it so NULL is passed into the deltacloud gahp if no
realm is specified by the user, allowing any realm to be chosen.

Based on a patch from Ian Main.

Signed-off-by: Chris Lalancette <clalance@redhat.com>