Condor Quota Check Classad Plugin
This patch creates a condor classad plugin that allows us to check
quotas as a final step in the condor matching process.

In order to enable quotas, you will have to add the following to your
/var/lib/condor/condor_local.config:

CLASSAD_USER_LIBS = /usr/share/deltacloud-aggregator/classad_plugin/deltacloud_classad_plugin.so

You must also set up the plugin correctly depending on how you plan to
run the aggregator.  The easiest thing to do is just use 'make rpms' and
install the resulting rpms and run the whole thing installed.

If you wish to run it from the src dir as most developers do, you will
have to run configure with:

./configure --enable-local --prefix=/usr

and then do a 'make install'.  This will install the
deltacloud_classad_plugin.so in
/usr/share/deltacloud-aggregator/classad_plugin directory, but it will
continue to use the database and model files from the local directory.

Signed-off-by: Ian Main <imain@redhat.com>