Fixes to the deltacloud classad quota plugin.
I could not get the previous version of the classad plugin
to work.  Somehow the full rails init (via require dutils)
was wreaking havoc with my condor instance.  Instead,
this patch just drags in the models that we need, and fixes
up the problem that way.  It also adds in additional logging
infrastructure so we have a better idea of what is going on.

Signed-off-by: Chris Lalancette <clalance@redhat.com>