Rename condor_refreshd.rb to condor_refreshd.
In addition, make it daemonize by default.  This makes
the whole thing much more daemon like, and makes it
appropriate for running from an initscript.

Speaking of which, the second part of this is to add a
new initscript to conf/ for this daemon.  With all of
this in place, things seem to be working pretty well.

Signed-off-by: Chris Lalancette <clalance@redhat.com>