Move condor classad sync'ing to condor_refreshd.
Always doing it in condor_refreshd has a number of benefits:

1)  It consolidates the logic in one place, getting rid of
possible race issues between two things doing it at the
same time.

2)  It makes it asynchrnous from the point of view of the
UI.  That means the UI just kicks the daemon, and then goes
on with life.  This should speed up the UI considerably.

3)  It makes it so that condor doesn't need to be running
to run the spec tests.

Signed-off-by: Chris Lalancette <clalance@redhat.com>