Add STATE_ERROR to the instance model and use it in dbomatic/condormatic.
This patch uses STATE_ERROR instead of STATE_CREATE_FAILED for the
instance since condor uses a generic 'held' state to indicate an error.
At this time we are not parsing out the error information so really we
are setting a generic error about the instance rather than the special
case error of CREATE_FAILED.

Obviously we need better error messages and states but for now this at
least more accurately reflects the nature of the state.

Signed-off-by: Ian Main <imain@redhat.com>