Lousy layout for the provider accounts page (3112)
This layout really doesn't work as a single row if input fields if we have 7 +
1 columns.  The horrible file input overflow could be worked around in JS, by
making it a button -- wrapping it in a relatively positioned element that
would hold the button styling, and making the input absolutely positioned,
transparent (will trigger clicks). But it still won't make this layout any
more usable.