Get error messages from condor into the db.
This patch teaches dbomatic how to get error messages into the database.
I added a 'last_error' field to the instance model so that we can stick
errors from condor in there.

Note that this is really not the right way to do this.  We should be using
the event logging system and tying the errors back to the instance from
that.  However this will work for now and give the users something
useful in case of an error.

Note that I haven't had a chance to test this patch live yet since my
setup is not working end to end yet.