Add classad sync after images are built.
This causes a classad sync to be called when images are done building.
This fixes a bug where new instances would error out for a while when
using a new image.

This version is updated to use kick_condor, and is only called when
a change occurs, not every iteration.

Signed-off-by: Ian Main <imain@redhat.com>