Do not add replicated images with nil image key
When an image build is requested, an entry is placed in the db for
a replicated image.  However it is not yet ready for use but we were
generating classads for them anyway.  This change makes it so we do
not generate classads unless an image key is present and they are
ready to be used.

Signed-off-by: Ian Main <imain@redhat.com>