Add a script to import some ec2 templates into the database.
These templates are "real" in the sense that you can launch
instances against them.  However, they are not advertised to
the iwhd, so they aren't the same as building the image through
the deltacloud UI.  Nonetheless, this script can be useful to
do testing on the deltacloud UI.

Script from Ian Main.

Signed-off-by: Ian Main <imain@redhat.com>
Signed-off-by: Chris Lalancette <clalance@redhat.com>