Make condor_refreshd refresh classads after 15 minutes.
In order to make some of the condor patches more palatable
to upstream, they requested that we stop using the phone-home
capability of condor to refresh our classads.  Since we don't
expect condor to restart very often, it should be sufficient to
just advertise classads to condor every 15 minutes or when it
is kicked.  This patch just puts the socket in a select with
a timeout and wakes up every 15 minutes to re-advertise.

Signed-off-by: Chris Lalancette <clalance@redhat.com>