BZ #640301
https://bugzilla.redhat.com/show_bug.cgi?id=640301

This allows the users to edit a URL of a provider that's been previously added
to the database.

Note that the backend provider must not change, this is only for cases when
the URL entry point changes. You can't use this to switch from EC2 to
Rackspace, for example.

Since actually swapping the provider backend would require some additional
work and possibly some user intervention, it would be better as separate
feature.