Remove condormatic_sync_instance_states.
This is no longer necessary now that dbomatic replays events.
It was also causing maintenance difficulties in that we would
have to parse the XML in two different places (once in dbomatic
and once in condormatic), which we were forgetting to do.
Just remove it completely, and rely on the replay to handle
events that we haven't consumed yet.

Signed-off-by: Chris Lalancette <clalance@redhat.com>