BZ#644534 Added check for Quota before creating instance
https://bugzilla.redhat.com/show_bug.cgi?id=644534

This patch adds a quota check to the webapp, so the user
will be notified that their instance cannot start until
they have free quota, but in the background condor will
continue to try to start the instance until such quota is
available. Instance will show in list as 'new' until
condor can do something with it.