Add the Resource Management -> Pools page
This is the first step towards implementing the Pools page in the new UI.

Actions `index` and `show` are in place and display the proper pages (though
the content is pretty much still missing).

The "New Pool" button is working; the "Edit" one is hooked up to the
Controller and has some basic code but the functionality is not there yet.