Updated HWP Matching, to return matched values, for non_defaults
Updated the HWP methods for matching Aoelus HWP to Provider HWPs.  This change now calculate the matching range and enum over laps as well as the matching HWP.  It also determines which hwp CONDOR should use, by the added match_hwp(hwp) method.

Logic for choosing a matched HWP and for choosing non_default values is very basic in this patch, at current it just returns the first in the list.