refactored role/privilege model
Privileges are now a combination of a base action and a target object type (i.e. "modify" on Pool instead of the flat string "pool_modify"). In addition, the modeling has changed -- instead of a privileges model that was only low-level privileges with an explicit mapping to roles, the privileges model now tracks the privileges assigned to each role, with no separate mapping table.

This will require a database reload as the seeds.rb required substantial changes in role/privilege loading

Signed-off-by: Scott Seago <sseago@redhat.com>