Add Solr/Sunspot search support
This gets us ready for adding the search. The Solr server is bundled with the
sunspot gem so you don't have to install it by hand. On the dev machine, not
in production of course.

Run `rake sunspot:solr:start` to start the Solr server.

When you make change to the solr/conf files, you must stop and start it again.

Run `rake sunspot:reindex` to rebuild the index -- do this when you change the
search setup in the models.