Disable quota plugin.
This is a TEMPORARY hack to get around the need to initialize rails in
the models.  The addition of search has made this plugin fail because
the pool and quota models now require rails to be initialized in order
to load.

The right way to fix this is to create a REST API allowing us to call
into the conductor for our quota check.

Signed-off-by: Ian Main <imain@redhat.com>