Task 197 - Remove any logic forcing one provider only
Inside this patch I rename model of CloudAccount to ProviderAccount. Also I refactor all codes where this model was called. All supported providers are included in constant hash, that is named PROVIDER_TYPES in Provider model. So now, attribute of class "cloud_type" is renamed to "provider_type" and it must be one of these PROVIDER_TYPES value. I fixed all Rspecs and Cucumbers for provider_account model.

In this revision added fixing failing rspec in next.

When you'll test this patch won't forget to run "rake db:migrate", because some of the attributes were changed in DB.