Crappy rpm workaround
Since 'make rpms' uses the existing directory to build an rpm from, it's
possible that it can pick up an existing condor plugin .so and install
it even if it's got the wrong prefix set etc.

This patch just adds a 'make clean' before 'make' so that it is forced
to rebuild a new plugin when making the rpm.  Clearly however the rpm
needs to be fixed.

Signed-off-by: Ian Main <imain@redhat.com>