Instance ssh key
When instance is started unique ssh key is generated and uploaded to instance.
New instance key is not generated with deltacloud api because api returns only
private part of key but with openssl lib.

This patch requires net_scp and delayed_job gems and delayedjob worker should be
started with: 'rake jobs:work' command