Task 433 and 434, template build ui changes
Renamed "Images" tab to "Builds"
Table columns now matches cloud-engine-templates-details-images.png mockup.
Table shows builds that have been completed and uploaded, and additional
providers that a completed build can be uploaded to. The additional provider's
provider type would have to match the build's target.

Note the Arch column may be redundant because the Architecture is specified
per template and not per build.

Added an additional column to select providers to upload an existing image.
The hook to upload the image will need to be done at a future sprint when the
new image factory is ready.

It also contains a new form element to initiate a build. The select drop down
will only show targets that haven't been built. Try creating a mock provider
and you should see it appear in the select.