Fix the failing cucumber tests
Certain cucumber tests were generating an admin user twice for a test (once in
the `Background` section and then again in the `Scenario`).

The previous commit changed factory definition for the admin user so that it
always has the login "admin". This caused failures for scenarios that tried to
create the admin user twice.

This removes the "Given I am an authorised user" condition from the `Scenario`
sections, because the admin user is already established in the `Background`
section.