Fix BZ 679147
https://bugzilla.redhat.com/show_bug.cgi?id=679147

In a JS-enabled browser, we use jQuery UI Tabs to display the contents of the
details pane. The content is loaded using ajax whenever the user switches a
tab.

The bug was that the first tab never got reloaded. That caused
inconsistencies in the UI.

As a part of the fix, the text in the tab headers is now enclosed in <span>
tags which allows us to show a message notifying the user that the tab is
being loaded.

For whatever reason, jQuery UI requires that <span> tag for this to work.