BZ 688924 - Get ec2 image pushes working.
https://bugzilla.redhat.com/show_bug.cgi?id=688924

Typhoeus was sending a Header that indicated it needed
permission from the server to continue sending the request
(Expect: 100-continue).  Thanks to Jan for helping discover
that detail.  For now, figuring out how to make the server
respond properly to that header was far more work than just
using a different client library (which we had as a dep for
deltacloud and other things already anyway). This patch just
switches the client and adds some tests to make sure it works.