Make ProviderAccount.build_credentials dynamic
This fixes Redmine #692

We're building the credentials XML that's passed to Image Factory dynamically
using the credential_definitions table now.

The advantage is that as we add new providers, we won't have to update the
build_credentials function. We populate the credential_definitions with the
new entries and that will be it.

Fix according to jay's suggestions

Fix the tests