Big rewrite of the aeolus-conductor spec file.
The idea here is to vastly clean up the spec file and
only pull in the pieces we care about.  To that end,
we no longer copy all of the src/ and then remove specific
bits.  Instead, we only copy over the bits we specifically
care about.  This should make the spec file a little more
sane.

Signed-off-by: Chris Lalancette <clalance@redhat.com>