Validate pools deletion
This fixes Redmine #1098

Pools can be deleted only if they either don't have any instances associated
with them or if all those instances are stopped and stateless.

This also provides a dummy check to see whether an instance is stateful or
stateless. Currently, we only support stateless instances, but when we modify
the instance.restartable? method later on, nothing should break.