Re-order the refreshd loop.
As it stands, if condormatic_classads_sync throws an exception,
it will just continue failing in a fast loop, filling up the
logs.  To prevent that, re-order the loop.  First, do a
classads_sync when we first come up; if it fails, continue
on anyway because later ones might succeed.  Next, take the
select out of the begin..rescue block; if select fails, there
is not a whole lot we can do and we should exit.  Finally,
but condormatic_classads_sync at the bottom of the loop.
That way, if it throws an exception, we'll go back to the
top of the loop and do another sleep on select.

Signed-off-by: Chris Lalancette <clalance@redhat.com>