BZ695770 / Task 1173 - instance creation gui needs redhat/fedora logos
Added logos for fedora and rhel. The image url uses tpl.platform which
includes the version number. May be worth refactoring that to drop the
version number. At the moment there are symlinks for each version
pointing to the single fedora logo.

Display of RHEL logo is untested. RHEL image imports sets the
tpl.platform value to "unknown". As a temporary measure, I've added
a conditional to not display an image if the value is "unknown".