Issue #1432: Make connector able to point at different conductor location.
https://www.aeolusproject.org/redmine/issues/1432

Because we want this to work with both the default sample config file
and allow the user or admin to use a file of their choosing, this patch
adds a bit of rework around parsing off options - specifically, it now
actually parses options with OptionParser.  This means the user can pass
in our new custom option to point to the config file, we can set defaults
on the options we care about, and thin can parse the rest.  The only caveat
here is that the command (start,stop, etc) must be passed in first, all
actual options can be passed in any order.