Remove condor_refreshd.
As pointed out by Mark McLoughlin, the whole system for
doing matching was much too complicated.  Matching is a
pretty simple problem, so instead of letting condor do
it, just do it inline in the conductor.  We still use
condor to launch and monitor the jobs.

Doing this saves us 750 lines of code, allows us to get rid
of condor_refreshd, allows us to get rid of the quota plugin,
and lets us get rid of autotools.  This makes the whole
system much simpler and more reliable.

This version removes password file support and fixes rpm
version (0.2.0 -> 0.3.0) so rpms upgrade properly - Ian.

Signed-off-by: Chris Lalancette <clalance@redhat.com>