Change condormatic to use DeltacloudPasswordFile.
Later versions of condor no longer have a password in
the classads, since this was a security issue.  Instead,
they use DeltacloudPasswordFile to specify a file to be
read which contains the password.

This patch changes our condormatic_instance_create code
to use this new DeltacloudPasswordFile support.  It does
this by writing a file in /var/lib/aeolus/conductor/jobs
containing the password, and then passing this filename
into condor.

Signed-off-by: Chris Lalancette <clalance@redhat.com>