Connect the Pretty/Filter toggle to viewstate
The user-selected view is stored in Viewstate and remembered in the
nav_history bar.

For instance if you select the Filter view in the Pools page, then drill down
to a single Pool view and then go back to Pools using the breadcrumbs link,
you'll end up with the Filter view, not Pretty view.