Add delete confirmation to Pools filter view pages
This is somewhat related to the bug #717538

https://bugzilla.redhat.com/show_bug.cgi?id=717538

When a table of objects was loaded via ajax (for example by switching to the
Filter View), the javascript click handlers set up in `multiDestroyValidation`
under application.js were lost.

As a result, some Delete buttons no longer asked for confirmation on delete.