Fix compatibility with the new Deltacloud client
The latest update of the Deltacloud client in Fedora made a change that broke
our current code:

When getting a Hardware Profile Property that a given Hardware Profile object
didn't have, the old client returned `nil` while the new one raises
`NoMethodError`.

Example of where this happens is the Mock driver. It has a Hardware Profile
called "Opaque" which doesn't have any properties defined.

This change of behaviour caused our test to fail and made it impossible to add
the Mock provider via the UI.

This fixes the Conductor code so that it's compatible with both versions of
the Deltacloud client. Later we may want to deprecate the older clients and
simplify the code somewhat.