final change for Bug 714790
Most of Bug 714790 has been pushed, but the remaining issue was that non-admin users, when clicking the 'administer' tab would result in an 'insufficient privileges' error.

As it turns out, the default action for 'administer' was the 'users' list, which happened to be the only 'administer' action that doesn't allow non-admin users from viewing the page at all. In other cases, normal users should get viewing (but not editing) rights -- HWPs, Realms, and Suggested deployables. In some, (providers, etc) the action is valid but the results list is filtered by permissions.

This is only a temporary fix, since UX design for 'administer' hasn't been done yet. I'm changing the default action to view 'Suggested Deployables' instead of 'Users', as everyone can see that. In addition, the nav pulldown filters 'Users' section is now conditionally included based on the same permissions check that complains if non-admin users attempt to navigate there. I'm not enforcing permissions in nav anywhere else, since the pages themselves don't enforce them.