Rename rhev-m/vmware target names to rhevm/vsphere
In the case of RHEV-M and VMWare, Conductor provider codenames differ
from deltacloud driver names. These different names were chosen because
they are what image factory uses.

However, in future, this will cause problems for conductor when we use
a single deltacloud instance because we will need to map our codenames
to deltacloud driver names.

We should take the opportunity and cut out all confusion by
standardising on deltacloud driver names. Obviously, we should only
do this in Conductor if the change is also made in image factory.