Ensure that the Quota association is never nil
The objects that used quota were valid even when the quota was nil. This added
additional complexity to the code and created a semantic confusion (what does
it mean that pool.quota is nil? And how does it differ from
pool.quota.maximum_running_instances being nil?).

This commit adds tests and validation in place to make sure that the quota
association will no longer be allowed to be nil.

Since the code always created the quota association anyway, no migrations
should be necessary.