When deleting a deployment, stop running instances
https://bugzilla.redhat.com/show_bug.cgi?id=721351

If all deployment's instances are either stopped or running, the UI will allow
the deployment to be deleted.

All the running instances will be stopped. Once that's done, the deployment
and all its associated instances will be deleted from the database.

Since we don't have a better way to handle arbitrary Condor callbacks, this
relies on dbomatic to set the instance to 'stopped' and the ImageObserver
'after_save' callback to trigger the deletion.