Use port 3002 for the Mock Deltacloud server
This tends to trip people up: aeolus-configure sets up a Deltacloud server
with the Mock driver on the port 3002.

However, our tests look for the mock dc instance on 3001, which means that to
run the tests, one must start another Mock Deltacloud server.

This makes tests use 3002 instead. Therefore, if you have the Conductor set up
for production (e.g. installed from RPMs), the tests will use the Deltacloud
server that came with it.

Otherwise, you'll still have to start it manually -- listening on the port
3002 from now on.