Update FactoryGirl use to current practice.
Factory girl recommends the use of FactoryGirl rather Factory and
FactoryGirl.create.build rather than simply Factory.

Also fixed some broken factory declarations. Either due to a bug or an feature,
Factory girl won't resolve sequences in dependent attributes. To work around
this, I placed the generation of the dependent attributes inside an after_build
block. I've also mailed the FG list to see if it is a bug, or whether it is
actually intended behaviour.