Seed data truncated at the end of each suite rather than each test.
We were loading seed data at the beginning of each suite then truncating at the
end of each test. This leaves the database with no seed data for all but the
first test in a suite.

Changed this so that seed data is only truncated after each suite to match. The
tests are all transactional anyway so each test is still isolated.