Remove database cleaner for specs and rely on db:test:prepare rake tast.
A full build always db:test:prepare so I've hooked in a call to load the seed
data at this point. This means we only need to load seed data once at the start
of a build thereby improving the speed. Moreover, each test is run within its
own transaction so the data will remain effectively untouched at the end of
each.

The real impact of this change is that any developer wishing to run a single
test now needs to ensure the test database is populated with seed data either
by running db:seed directly, or via db:test:prepare. I don't see this as an
issue as it's pretty common practice to ensure the test database is prepared
correctly after pulling changes anyway.