Fix the login page error reporting
https://bugzilla.redhat.com/show_bug.cgi?id=726779

This handles two issues with our login-over-javascript situation.

The first was a broken UI when the error message was unexpectedly a HTML page
instead of a single text response that we set on entering bad credentials.

This could happen if the Rails app or Apache returned a 5xx error when the
user tried to log in.

The second was not being able to get through the login page if the user has
already signed in in another tab.