Reenabled restful_submit
Not sure what's the plan about this thing but at this point we
use it on many places. If we don't want to enable it, we have to
update all forms where it's used and suppose that we have only one
action per form (or use dispatchers as before). I think that decision
if we keep this or not depends on UI design, until we have it
reenabling this plugin is simplest solution.