Update dbomatic for the new Nokogiri interface
The Nokogiri that's in Fedora 15 has a different interface to its
`start_element` SAX method.

The `attributes` argument is a list of tuples where each tuple contains the
`name` and `value` of an attribute. Previously, it was a flat array of
strings.

This patch makes it work both for Nokogiri 1.4.x and 1.5.x.