added settings.yml with default deltacloud URL
With the change to a single deltacloud-core provider in -configure (not yet in place), we want to default the url parameter for new providers to this value. A user adding a new provider won't have to know where this is -- it's filled in by default in the form. The user can change it, of  course.