Disable Backbone fallback to hash-based URLs
In browsers that don't support the HTML5 History API, Backbone redirects URLs
to the hash-based format.

E.g. /deployments/1 becomes /#deployments/1

This behaviour breaks the app, because Rails render the page located at the
root path.

Since we don't use the Backbone's capabilities that require this behaviour, we
disable it for the time being.

If we want to use them later, a modification of Rails routing logic will be
required.